address comments on the pr
